+++ src/model/scure-rooms-model.js
@@ -9,16 +9,17 @@ const getUnlockedIds = unlocked => (destination) => {
   if (isUnlocked(unlocked, destination)) return destination.roomId;
   return null;
 };
-const getLockedIds = unlocked => (destination) => {
-  if (destination.isLockedDestination && !isUnlocked(unlocked, destination)) return destination.roomId;
-  return null;
-};
+const getLockedIds = unlocked => destination =>
+  ((destination.isLockedDestination && !isUnlocked(unlocked, destination)) ?
+    destination.roomId : null);
 const notNull = d => d !== null;
 const getUnlockedDestinationsIds = (theMap, fromId, unlocked) =>
   theMap[fromId].map(getUnlockedIds(unlocked)).filter(notNull);
 const getLockedDestinationIds = (theMap, fromId, unlocked) =>
   theMap[fromId].map(getLockedIds(unlocked)).filter(notNull);
-const roomById = id => r => (r.isLockedDestination && r.roomId === id) || (!r.isLockedDestination && r === id);
+const roomById = id => r =>
+  (r.isLockedDestination && r.roomId === id) || (!r.isLockedDestination && r === id);
+const byLockedDestinationAndId = roomId => r => r.isLockedDestination && r.roomId === roomId;
 const isADestination = (theMap, fromId, roomId) => theMap[fromId].find(roomById(roomId));
 const isALockedDestination = (theMap, fromId, roomId, unlocked) =>
   getLockedDestinationIds(theMap, fromId, unlocked).indexOf(roomId) >= 0;
@@ -52,7 +53,7 @@ class ScureRoomsModel {
 
   getDestinationLockedSentence(destinationName, fromRoomId) {
     const room = this.getRoomByName(destinationName);
-    const theLockedDestination = this.map[fromRoomId].find(r => r.isLockedDestination && r.roomId === room.id);
+    const theLockedDestination = this.map[fromRoomId].find(byLockedDestinationAndId(room.id));
     return theLockedDestination ? theLockedDestination.sentence : null;
   }