+++ test/common.js
@@ -1,9 +1,9 @@
 global.chai = require('chai');
 global.sinon = require('sinon');
+const { buildScureFor } = require('../src/scure.js');
+const { data } = require('./data/data-test');
+
 global.chai.should();
 global.expect = global.chai.expect;
 
-global.buildTestScure = () => {
-  const { data } = require('./data/data-test');
-  return require('../src/scure.js').buildScureFor(data);
-};
+global.buildTestScure = () => buildScureFor(data);